[TrivialFix] Moved floodlight exceptions to 'error' package
 - Exception classes were moved to 'error' package
 - Fixed some checkstyle issues
