--- services/src/floodlight-modules/src/main/java/org/openkilda/floodlight/switchmanager/SwitchNotFoundException.java
@@ -1,9 +1,0 @@
-package org.openkilda.floodlight.switchmanager;
-
-import org.projectfloodlight.openflow.types.DatapathId;
-
-public class SwitchNotFoundException extends SwitchOperationException {
-    public SwitchNotFoundException(DatapathId dpId) {
-        super(dpId, String.format("Switch %s was not found", dpId));
-    }
-}
