--- services/src/floodlight-modules/src/main/java/org/openkilda/floodlight/switchmanager/SwitchOperationException.java
@@ -1,20 +1,0 @@
-package org.openkilda.floodlight.switchmanager;
-
-import org.projectfloodlight.openflow.types.DatapathId;
-
-public class SwitchOperationException extends Exception {
-    private final DatapathId dpId;
-
-    public SwitchOperationException(DatapathId dpId) {
-        this(dpId, "Switch manipulation has failed");
-    }
-
-    public SwitchOperationException(DatapathId dpId, String message) {
-        super(message);
-        this.dpId = dpId;
-    }
-
-    public DatapathId getDpId() {
-        return dpId;
-    }
-}
