+++ services/src/floodlight-modules/src/main/java/org/openkilda/floodlight/switchmanager/web/FlowsResource.java
@@ -17,8 +17,8 @@ package org.openkilda.floodlight.switchmanager.web;
 
 import static org.openkilda.messaging.Utils.MAPPER;
 
+import org.openkilda.floodlight.error.SwitchOperationException;
 import org.openkilda.floodlight.switchmanager.ISwitchManager;
-import org.openkilda.floodlight.switchmanager.SwitchOperationException;
 import org.openkilda.floodlight.utils.CorrelationContext;
 import org.openkilda.messaging.error.ErrorType;
 import org.openkilda.messaging.error.MessageError;
